Fix Check Ready Race Condition and ExecCmd Thread Lock (#268)
Fix Check Ready Race Condition and ExecCmd Thread Lock

The client CheckReady command could be called with zero apps loaded in some cases which can lead to the port forwarding failing intermittently. This changes it to wait for at least 1 app to show up before starting the check. This should fix most cases but we may still want a more comprehensive check that waits for the expected number of apps instead.

The client ExecCmd could thread lock on in some rare cases due to the std out blocking the stderr from finishing. Moving the scanning of both the stdout and stderr to their own go routines prevents this from happening.